Refactored paf_to_m8 to move it to paf.rs.
